Moved the evaluation-related functions to a seperated new class and documented code
